chore: update template
